requirements.txt
Pillow==10.1.0
PyQt5==5.15.9
keyboard==0.13.5
requests==2.31.0
gradio==4.44.1
